+++ plugins/form/CHANGELOG.md
@@ -1,3 +1,21 @@
+# v2.0.5
+## 09/15/2016
+
+1. [](#bugfix)
+    * Fix passing updating the header through event, no need for return value
+
+# v2.0.4
+## 09/15/2016
+
+1. [](#improved)
+    * Allow filling the page header form dynamically (e.g. use case: Comments plugin)
+
+# v2.0.3
+## 09/12/2016
+
+1. [](#improved)
+    * Use `Page::slug()` for form name if not set in the form itself (better backwards compatibility)
+
 # v2.0.2
 ## 09/08/2016